Error Handling RenderDoc Module
I added Error-Handling for the import of the RenderDoc Module in google_maps_rd.py

It now will give more details on the console level when something doesn't work as expected.
And I also added more specific error messages to apear on the GUI when it's not possible to import the RenderDoc module.